requirements.txt
discord.py==2.3.2
ffmpeg-python
pynacl
opuslib
